storybook - organize story groups in directories
