+++ .gitignore
@@ -11,6 +11,9 @@
 # production
 /dist
 
+# IDE
+.vscode
+
 # misc
 .DS_Store
 .env.local
