+++ stories/nivo/bar.stories.js
@@ -1,17 +1,17 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import BarChart from '../src/components/bar-chart/'
-import barChartData from './data/others/bar-chart-data'
-import bannerData from './data/atom/overlord-banner'
-import scatterChartData from './data/others/scatter-chart-data'
-import lineChartData from './data/others/line-chart-data'
+import BarChart from '../../src/components/bar-chart/'
+import barChartData from '../data/others/bar-chart-data'
+import bannerData from '../data/atom/overlord-banner'
+import scatterChartData from '../data/others/scatter-chart-data'
+import lineChartData from '../data/others/line-chart-data'
 import ResponsiveChartWrapper from './responsive-chart-wrapper'
 
 
 const title = 'My Title'
 
-storiesOf('Bar Chart', module)
+storiesOf('Nivo/Bar Chart', module)
   .add('Stacked', () => (
     <ResponsiveChartWrapper>
       <BarChart
