+++ stories/nivo/title.stories.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import ScatterChart from '../src/components/scatter-chart'
-import scatterChartData from './data/others/scatter-chart-data'
+import ScatterChart from '../../src/components/scatter-chart'
+import scatterChartData from '../data/others/scatter-chart-data'
 import ResponsiveChartWrapper from './responsive-chart-wrapper'
 
 
 const title = 'My Title'
 
-storiesOf('Title', module)
+storiesOf('Nivo/Title', module)
   .add('Widget Scatter Chart with no title', () => (
     <ResponsiveChartWrapper>
       <ScatterChart
