+++ stories/nivo/tooltip.stories.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import PieChart from '../src/components/pie-chart'
-import pieChartData from './data/others/pie-chart-data'
+import PieChart from '../../src/components/pie-chart'
+import pieChartData from '../data/others/pie-chart-data'
 import ResponsiveChartWrapper from './responsive-chart-wrapper'
 
 
 const title = 'My Title'
 
-storiesOf('Tooltip', module)
+storiesOf('Nivo/Tooltip', module)
   .add('Tooltip Format', () => (
     <ResponsiveChartWrapper>
       <PieChart title={title} data={pieChartData} isDonut={false} tooltipFormat={v => `$${v}`} />
